refactor frc.robot.extras.simulation directory
