+++ src/main/java/frc/robot/extras/simulation/field/SimulatedField.java
@@ -1,10 +1,13 @@
-package frc.robot.extras.simulation;
+package frc.robot.extras.simulation.field;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.extras.simulation.physicsSim.AbstractDriveTrainSimulation;
+import frc.robot.extras.simulation.gamePiece.CrescendoNoteSimulation;
+import frc.robot.extras.simulation.gamePiece.GamePieceSimulation;
+import frc.robot.extras.simulation.mechanismSim.IntakeSimulation;
+import frc.robot.extras.simulation.mechanismSim.swervePhysicsSim.AbstractDriveTrainSimulation;
 import frc.robot.extras.util.GeomUtil;
 import java.util.*;
 import org.dyn4j.dynamics.Body;
@@ -174,7 +177,7 @@ public abstract class SimulatedField {
    * part of its collision space.
    *
    * <p>This method immediately starts the {@link
-   * org.ironmaple.simulation.IntakeSimulation.GamePieceContactListener}, which listens for contact
+   * frc.robot.extras.simulation.mechanismSim.ironmaple.simulation.IntakeSimulation.GamePieceContactListener}, which listens for contact
    * between the intake and any game piece.
    *
    * @param intakeSimulation the intake simulation to be registered
