+++ src/main/java/frc/robot/extras/simulation/gamePiece/CrescendoNoteSimulation.java
@@ -1,7 +1,8 @@
-package frc.robot.extras.simulation;
+package frc.robot.extras.simulation.gamePiece;
 
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.util.Units;
+
 import org.dyn4j.geometry.Geometry;
 
 public class CrescendoNoteSimulation extends GamePieceSimulation {
