+++ src/main/java/frc/robot/subsystems/swerve/moduleIO/SimulatedModule.java
@@ -6,7 +6,8 @@ import edu.wpi.first.math.controller.SimpleMotorFeedforward;
 import edu.wpi.first.math.trajectory.TrapezoidProfile.Constraints;
 import edu.wpi.first.math.util.Units;
 import frc.robot.extras.simulation.OdometryTimestampsSim;
-import frc.robot.extras.simulation.physicsSim.SwerveModuleSimulation;
+import frc.robot.extras.simulation.mechanismSim.swervePhysicsSim.SwerveModuleSimulation;
+
 import java.util.Arrays;
 
 /** Wrapper class around {@link SwerveModuleSimulation} that implements ModuleIO */
